# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management System",
    "summary": "Multi-tenant school management platform",
    "description": "EduCloud API -- identity, tenancy, RBAC, plans and invitations.\n\n**Tenant isolation** is enforced by PostgreSQL Row-Level Security. The `org` claim in your access token scopes every query at the database level; cross-tenant reads return 404, never 403, so resource existence is never disclosed.\n\n**School scoping** is a second, softer boundary inside the organization: a school-scoped membership sees only its own campus, while an organization owner spans all of them.\n\n**Permissions** are `resource:action` strings resolved per request from the caller's role. They are deliberately not embedded in the token, so revoking access takes effect on the very next request.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/students/admissions": {
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Submit a public admissions application",
        "description": "Public endpoint backing the Next.js admissions form.\n\n`PublicDbSession`, not `DbSession`: there is no caller identity. Declared before\n`/{student_id}` so \"admissions\" is never parsed as a student id.",
        "operationId": "submit_admission_api_v1_students_admissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentAdmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdmissionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/students": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "List and search students",
        "operationId": "list_students_api_v1_students_get",
        "parameters": [
          {
            "name": "section_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by section.",
              "title": "Section Id"
            },
            "description": "Filter by section."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/StudentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by enrollment status, e.g. `pending` for the admissions queue.",
              "title": "Status"
            },
            "description": "Filter by enrollment status, e.g. `pending` for the admissions queue."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort By"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDirection",
              "default": "asc"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term",
              "title": "Q"
            },
            "description": "Search term"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StudentRead_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "post": {
        "tags": [
          "Students"
        ],
        "summary": "Enroll a student",
        "operationId": "create_student_api_v1_students_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/students/{student_id}": {
      "get": {
        "tags": [
          "Students"
        ],
        "summary": "Get a student",
        "operationId": "get_student_api_v1_students__student_id__get",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Students"
        ],
        "summary": "Update a student",
        "operationId": "update_student_api_v1_students__student_id__patch",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudentRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Students"
        ],
        "summary": "Remove a student from the directory",
        "description": "Soft delete -- fee, attendance and certificate history is preserved.",
        "operationId": "delete_student_api_v1_students__student_id__delete",
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Student Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AdmissionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "admission_number": {
            "type": "string",
            "title": "Admission Number"
          },
          "status": {
            "$ref": "#/components/schemas/StudentStatus"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": "Application received. The school will contact you."
          }
        },
        "type": "object",
        "required": [
          "id",
          "admission_number",
          "status"
        ],
        "title": "AdmissionResponse",
        "description": "Deliberately thin. The public form must not echo back the stored record --\nthat would turn the endpoint into a way to probe another school's data."
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other"
        ],
        "title": "Gender"
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          }
        },
        "type": "object",
        "required": [
          "page",
          "size",
          "total",
          "pages",
          "has_next",
          "has_prev"
        ],
        "title": "PageMeta",
        "description": "Pagination metadata; everything a table UI needs to render its controls."
      },
...
      "Page_StudentRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudentRead"
            },
            "type": "array",
            "title": "Items"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "meta"
        ],
        "title": "Page[StudentRead]"
      },
...
      "SortDirection": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection"
      },
      "StudentAdmissionRequest": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name"
          },
          "guardian_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Phone"
          },
          "guardian_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Email"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "school_id": {
            "type": "string",
            "format": "uuid",
            "title": "School Id",
            "description": "Which school is being applied to."
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "school_id"
        ],
        "title": "StudentAdmissionRequest",
        "description": "Public admissions form payload (PDF: \"Digital Admissions Form\").\n\nHas no `admission_number`, no `section_id` and no `status`: an applicant cannot\nassign themselves a roll number, place themselves in a class, or admit\nthemselves. The service generates the number and forces status to PENDING."
      },
      "StudentCreate": {
        "properties": {
          "first_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name"
          },
          "guardian_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Phone"
          },
          "guardian_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Email"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "admission_number": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Admission Number",
            "examples": [
              "2026-001"
            ]
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "status": {
            "$ref": "#/components/schemas/StudentStatus",
            "default": "active"
          },
          "enrolled_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrolled On"
          }
        },
        "type": "object",
        "required": [
          "first_name",
          "last_name",
          "admission_number"
        ],
        "title": "StudentCreate",
        "description": "Staff-facing creation: the student is enrolled immediately."
      },
      "StudentRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "admission_number": {
            "type": "string",
            "title": "Admission Number"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name"
          },
          "guardian_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Phone"
          },
          "guardian_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Email"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "status": {
            "$ref": "#/components/schemas/StudentStatus"
          },
          "enrolled_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrolled On"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "admission_number",
          "first_name",
          "last_name",
          "full_name",
          "date_of_birth",
          "gender",
          "address",
          "photo_url",
          "guardian_name",
          "guardian_phone",
          "guardian_email",
          "emergency_contact_name",
          "emergency_contact_phone",
          "section_id",
          "status",
          "enrolled_on",
          "created_at",
          "updated_at"
        ],
        "title": "StudentRead"
      },
      "StudentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "inactive",
          "graduated",
          "transferred"
        ],
        "title": "StudentStatus",
        "description": "Enrollment lifecycle.\n\nPENDING is what makes the PDF's \"Digital Admissions Form\" work: a public\napplication lands as PENDING and appears in an admissions queue, without ever\ncounting as an enrolled student or appearing in a class register."
      },
      "StudentUpdate": {
        "properties": {
          "first_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "gender": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Gender"
              },
              {
                "type": "null"
              }
            ]
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "photo_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Url"
          },
          "guardian_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Name"
          },
          "guardian_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Phone"
          },
          "guardian_email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guardian Email"
          },
          "emergency_contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Name"
          },
          "emergency_contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Contact Phone"
          },
          "section_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Id"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "enrolled_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrolled On"
          }
        },
        "type": "object",
        "title": "StudentUpdate",
        "description": "PATCH: every field optional, omission means \"leave unchanged\"."
      },
...
    }
  }
}